+++ multimodal/src/autogluon/multimodal/learners/__init__.py
@@ -1,3 +1,4 @@
 from .base import BaseLearner
+from .matching import MultiModalMatcher
 from .ner import NERLearner
 from .object_detection import ObjectDetectionLearner
